ft: metadata import capability in ObjectMD
ObjectMD model class is now able to import metadata from stored blob
and convert it to the latest version internally.

The first use case will be for backbeat to handle metadata checks and
updates, but it should also give a cleaner way to import and
manipulate object metadata in S3.

Note: removed 'version' parameter from ObjectMD constructor since it's
handled internally by the class and should not be exposed anymore.